matchmaking/requirements.txt
pika==1.3.0
dataclasses-json==0.5.7
cassandra-driver==3.25.0
redis==4.3.4